requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2
django-rest-passwordreset
